refactor(cli): modify cli code to be more readable and granual
