+++ packages/cli/index.ts
@@ -2,34 +2,22 @@
 import fs from 'fs';
 import chalk from 'chalk';
 import { Command } from 'commander';
-import gradient from 'gradient-string';
 import inquirer from 'inquirer';
 import { createProject } from 'stplr-core';
 import { checkAuthentication } from './utils/checkAuthentication';
 import { checkTools } from './utils/checkTools';
-import { findUnfinishedProjects, UnfinishedProject } from './utils/findUnfinishedProjects';
+import { findUnfinishedProjects, getProjectChoices, UnfinishedProject } from './utils/findUnfinishedProjects';
+import { unfinishedProjectsChoice } from './command-prompts/unfinishedProjectsChoicePrompt';
+import { overwriteDirectoryPrompt } from './command-prompts/overwriteDirectoryPrompt';
+import { shouldUsePayloadPrompt } from './command-prompts/shouldUsePayloadPrompt';
+import { displayHeader } from './utils/displayHeader';
+import { getProjectNamePrompt } from './command-prompts/getProjectNamePrompt';
 
-const asciiArt = `
-.&&&%                                                         &&&&                                    
-.&&&%                                                         &&&&                                    
-.&&&&&&&&*    (&&&&&&&&&&*      (&&&.&&&&&&&&&&      *&&&#    &&&&     #&&&&,                         
-.&&&&((((,  %&&&&(, .,#&&&&#    (&&&&&%(**(%&&&&(    *&&&#    &&&&   (&&&%                            
-.&&&%      %&&&.        ,&&&%   (&&&/        &&&&.   *&&&#    &&&& #&&&#                              
-.&&&%     ,&&&*          (&&&.  (&&&/        %&&&,   *&&&#    &&&&&&&&&&.                             
-.&&&%      %&&&.        *&&&#   (&&&/        %&&&,   *&&&#    &&&&&* *&&&%                            
-.&&&%       %&&&&%.  ,&&&&&#    (&&&/        %&&&,   *&&&#    &&&&     #&&&/                          
-.&&&%         (&&&&&&&&&%*      (&&&/        %&&&,   *&&&#    &&&&      .&&&&,                        
-`;
-
-const displayHeader = () => {
-  const metalGradient = gradient([
-    { color: '#3C3C3C', pos: 0 },
-    { color: '#FFFFFF', pos: 1 },
-  ]);
-
-  console.log(metalGradient(asciiArt));
-  console.log(chalk.bold('\nWelcome to Stapler!\n'));
-};
+interface Flags {
+  local?: boolean;
+  name?: string;
+  skipPayload?: boolean;
+}
 
 const program = new Command();
 
@@ -51,53 +39,32 @@ program
 
 program.parse(process.argv);
 
-interface Flags {
-  local?: boolean;
-  name?: string;
-  skipPayload?: boolean;
-}
-
 const createAction = async (options: Flags) => {
   const shouldDeploy = !options.local;
   const currentDir = process.cwd();
-  const unfinishedProjects: UnfinishedProject[] = findUnfinishedProjects(currentDir);
-  let selectedProject: UnfinishedProject | null = null;
+
   let proceedWithNewProject = true;
+  let selectedProject: UnfinishedProject | null = null;
 
+  const unfinishedProjects: UnfinishedProject[] = findUnfinishedProjects(currentDir);
+
+  // If no project name is provided, and there are unfinished projects, we prompt the user to resume one of them
   if (!options.name && unfinishedProjects.length > 0) {
-    const projectChoices = unfinishedProjects.map((proj) => ({
-      name: proj.projectName,
-      value: proj,
-    }));
+    const projectChoices = getProjectChoices(unfinishedProjects);
 
-    const resumeAnswers = await inquirer.prompt([
-      {
-        type: 'confirm',
-        name: 'resume',
-        message: `We found the following unfinished project(s):\n${unfinishedProjects
-          .map((p) => `- ${p.projectName}`)
-          .join('\n')}\nWould you like to resume one of them?`,
-        default: true,
-      },
-      {
-        type: 'list',
-        name: 'selectedProject',
-        message: 'Select a project to resume:',
-        choices: projectChoices,
-        when: (answers) => answers.resume && unfinishedProjects.length > 1,
-      },
-    ]);
+    const { resume, unfinishedSelectedProject } = await unfinishedProjectsChoice(unfinishedProjects, projectChoices);
 
-    if (resumeAnswers.resume) {
+    if (resume) {
       proceedWithNewProject = false;
       if (unfinishedProjects.length === 1) {
         selectedProject = unfinishedProjects[0];
       } else {
-        selectedProject = resumeAnswers.selectedProject || null;
+        selectedProject = unfinishedSelectedProject || null;
       }
     }
   }
 
+  // resume selected project
   if (!proceedWithNewProject && selectedProject) {
     process.chdir(selectedProject.projectPath);
     selectedProject.state.options.name = selectedProject.projectName;
@@ -107,40 +74,25 @@ const createAction = async (options: Flags) => {
       await checkTools();
     }
 
-    await createProject(selectedProject.state.options, selectedProject.projectPath).catch((error) => {
-      console.error('Error resuming project:', error);
-    });
+    await createProject({ ...selectedProject.state.options, shouldDeploy }, selectedProject.projectPath).catch(
+      (error) => {
+        console.error('Error resuming project:', error);
+      },
+    );
   } else {
     // Create new project
     options.name &&
       console.log('You have provided a project name of:', chalk.yellow(options.name), "let's continue...");
 
-    const projectName =
-      options.name ||
-      (
-        await inquirer.prompt([
-          {
-            type: 'input',
-            name: 'name',
-            message: 'What is your project named?',
-            default: 'my-stapled-app',
-          },
-        ])
-      ).name;
+    const projectName = options.name || (await getProjectNamePrompt());
 
     const projectDir = `${currentDir}/${projectName}`;
-    // Check if the directory already exists
-    if (fs.existsSync(projectDir)) {
-      const overwriteAnswer = await inquirer.prompt([
-        {
-          type: 'confirm',
-          name: 'overwrite',
-          message: `The directory "${projectName}" already exists. Do you want to overwrite it?`,
-          default: false,
-        },
-      ]);
+    const directoryExists = fs.existsSync(projectDir);
 
-      if (!overwriteAnswer.overwrite) {
+    if (directoryExists) {
+      const shouldOverwrite = (await overwriteDirectoryPrompt(projectName)).overwrite;
+
+      if (shouldOverwrite === false) {
         console.log(chalk.red('Project creation canceled. Project directory already exists.'));
         return;
       }
@@ -151,16 +103,7 @@ const createAction = async (options: Flags) => {
     }
 
     // Skip Payload if specified by the flag
-    const payloadAnswer = options.skipPayload
-      ? { usePayload: false }
-      : await inquirer.prompt([
-          {
-            type: 'confirm',
-            name: 'usePayload',
-            message: 'Would you like to add Payload to your app?',
-            default: true,
-          },
-        ]);
+    const payloadAnswer = options.skipPayload ? { usePayload: false } : await shouldUsePayloadPrompt();
 
     const finalOptions = { name: projectName, shouldDeploy: !options.local, ...payloadAnswer };
     if (shouldDeploy) {
