+++ packages/cli/utils/displayHeader.ts
@@ -1,0 +1,24 @@
+import gradient from 'gradient-string';
+import chalk from 'chalk';
+
+const asciiArt = `
+.&&&%                                                         &&&&                                    
+.&&&%                                                         &&&&                                    
+.&&&&&&&&*    (&&&&&&&&&&*      (&&&.&&&&&&&&&&      *&&&#    &&&&     #&&&&,                         
+.&&&&((((,  %&&&&(, .,#&&&&#    (&&&&&%(**(%&&&&(    *&&&#    &&&&   (&&&%                            
+.&&&%      %&&&.        ,&&&%   (&&&/        &&&&.   *&&&#    &&&& #&&&#                              
+.&&&%     ,&&&*          (&&&.  (&&&/        %&&&,   *&&&#    &&&&&&&&&&.                             
+.&&&%      %&&&.        *&&&#   (&&&/        %&&&,   *&&&#    &&&&&* *&&&%                            
+.&&&%       %&&&&%.  ,&&&&&#    (&&&/        %&&&,   *&&&#    &&&&     #&&&/                          
+.&&&%         (&&&&&&&&&%*      (&&&/        %&&&,   *&&&#    &&&&      .&&&&,                        
+`;
+
+export const displayHeader = () => {
+  const metalGradient = gradient([
+    { color: '#3C3C3C', pos: 0 },
+    { color: '#FFFFFF', pos: 1 },
+  ]);
+
+  console.log(metalGradient(asciiArt));
+  console.log(chalk.bold('\nWelcome to Stapler!\n'));
+};
